requirements.txt
# dependencies
asyncio==3.4.3
beautifulsoup4==4.12.2
colorama==0.4.6
duckduckgo_search==3.9.8
md2pdf==1.0.1
openai~=1.3.3
playwright==1.39.0
python-dotenv~=1.0.0
pyyaml==6.0.1
selenium==4.15.2
webdriver-manager==4.0.1
uvicorn==0.24.0.post1
pydantic==2.5.1
fastapi==0.104.1
python-multipart==0.0.6
markdown==3.5.1
anyio==3.7.1
langchain==0.0.335
tavily-python==0.2.8
permchain==0.0.6
arxiv==2.0.0
PyMuPDF==1.23.6
requests==2.31.0
jinja2==3.1.2
google-cloud-pubsub==2.19.0
PyJWT==2.8.0